requirements.txt
paho-mqtt==1.6.1
Faker==13.2.0
influxdb-client==1.26.0
ciso8601==2.2.0
python-dotenv==0.19.2